+++ CHANGELOG.md
@@ -48,11 +48,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed `compas.data.Data.name` to be stored in `compas.data.Data.attributes`.
 * Changed `compas.data.Data.__jsondump__` to include `compas.data.Data.attributes` if the dict is not empty.
 * Changed `compas.data.Data.__jsonload__` to update `compas.data.Data.attributes` if the attribute dict is provided.
+* Changed `compas.datastructures.Graph` to take additional `**kwargs`, instead of only `name=None` specifically.
+* Changed `compas.datastructures.Network` to take additional `**kwargs`, instead of only `name=None` specifically.
+* Changed `compas.datastructures.Halfedge` to take additional `**kwargs`, instead of only `name=None` specifically.
+* Changed `compas.datastructures.Mesh` to take additional `**kwargs`, instead of only `name=None` specifically.
 
 ### Removed
 
 * Removed `compas_rhino.forms`. Forms will be moved to `compas_ui`.
 * Removed `compas.datastructures.Datastructure.attributes` and `compas.datastructures.Datastructure.name` (moved to `compas.data.Data`).
+* Removed `attributes` from `compas.datastructures.Graph.data`.
+* Removed `attributes` from `compas.datastructures.Network.data`.
+* Removed `attributes` from `compas.datastructures.Halfedge.data`.
+* Removed `attributes` from `compas.datastructures.Mesh.data`.
 
 ## [2.0.0-beta.1] 2023-12-20